Fix constructors for custom container
Previously I (wrongly) used containers for the construction of maps when
a custom container was specified. That's not really how it is intended,
the parameter should still be an allocator.

This fixes that issues, so now its possible to copy/move etc. maps when
a custom container is used. That container needs the constructors that
accept an allocator.

I had to bump version by a major number to 2.0.0 because this is an
incompatible API change.