segment_map: Allow also specifying a custom container for the bucket array
The previous patch allowed using the bucket array in segmented mode.

This patch expands the series to similarly also allow using a custom
container.

We are only allowing specifying a custom container and derive a possible
custom allocator directly by rebinding the value allocator. This
guarantees that they can be converted to each other.